(MNIST) Align with 2.0.0

+++ mnist/src/mnist.h
@@ -46,7 +46,7 @@ public:
 	MNIST();
 
 	/// Inherited via LearningEnvironment
-	virtual void doAction(uint64_t actionID) override;
+	virtual void doAction(double actionID) override;
 
 	/// Inherited via LearningEnvironment
 	virtual void reset(size_t seed = 0, Learn::LearningMode mode = Learn::LearningMode::TRAINING,
